requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
django-filter
python-dotenv
psycopg2-binary
